+++ src/Plugin.cs
@@ -24,6 +24,7 @@ namespace LotusEcarlateChanges;
 [BepInDependency("fall_damage_for_creatures", BepInDependency.DependencyFlags.HardDependency)]
 [BepInDependency("goldenrevolver.CapeAndTorchResistanceChanges", BepInDependency.DependencyFlags.HardDependency)]
 [BepInDependency("MonsterLabZ", BepInDependency.DependencyFlags.HardDependency)]
+[BepInDependency("neobotics.valheim_mod.maxaxe", BepInDependency.DependencyFlags.HardDependency)]
 [BepInDependency("nex.SpeedyPaths", BepInDependency.DependencyFlags.HardDependency)]
 [BepInDependency("org.bepinex.plugins.backpacks", BepInDependency.DependencyFlags.HardDependency)]
 [BepInDependency("southsil.SouthsilArmor", BepInDependency.DependencyFlags.HardDependency)]
@@ -69,6 +70,7 @@ public class Plugin : BaseUnityPlugin
       new CustomKeyHints(),
       new FallDamageForCreatures(),
       new Lightsources(),
+      new MaxAxe(),
       new RenegadeVikings(),
       new SpeedyPaths(),
       new Tutorials(),